third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-first-valid-applies-expected.txt

This is a testharness.js-based test.
FAIL Tree order decides which meta color-scheme applies. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "dark"
Harness: the test ran to completion.
